+++ api/controller/unityController.js
@@ -1,4 +1,5 @@
 Unity = require("../model/unityModel");
+Register = require("../model/registerModel");
 functions = require("../utils/functions");
 
 exports.index = function (_, res) {
@@ -74,13 +75,23 @@ exports.update = function (req, res) {
 };
 
 exports.delete = function (req, res) {
-  Unity.deleteOne(
-    {
-      _id: req.params.unity_id,
-    },
-    function (err, _) {
-      if (err) res.send(err);
-      else res.json({ status: "ok", message: "unity deleted" });
+  Register.find({ vaccine_id: req.params.unity_id }, function (_, register) {
+    if (register) {
+      res.statusCode = 400;
+      res.json({
+        status: "error",
+        message: "can't delete because this unity is in use on registers.",
+      });
+    } else {
+      Unity.deleteOne(
+        {
+          _id: req.params.unity_id,
+        },
+        function (err, _) {
+          if (err) res.send(err);
+          else res.json({ status: "ok", message: "unity deleted" });
+        }
+      );
     }
-  );
+  });
 };
