+++ api/controller/vaccineController.js
@@ -1,4 +1,5 @@
 Vaccine = require("../model/vaccineModel");
+Register = require("../model/registerModel");
 functions = require("../utils/functions");
 
 exports.index = function (_, res) {
@@ -72,13 +73,23 @@ exports.update = function (req, res) {
 };
 
 exports.delete = function (req, res) {
-  Vaccine.deleteOne(
-    {
-      _id: req.params.vaccine_id,
-    },
-    function (err, _) {
-      if (err) res.send(err);
-      else res.json({ status: "ok", message: "vaccine deleted" });
+  Register.find({ vaccine_id: req.params.vaccine_id }, function (_, register) {
+    if (register) {
+      res.statusCode = 400;
+      res.json({
+        status: "error",
+        message: "can't delete because this vaccine is in use on registers.",
+      });
+    } else {
+      Vaccine.deleteOne(
+        {
+          _id: req.params.vaccine_id,
+        },
+        function (err, _) {
+          if (err) res.send(err);
+          else res.json({ status: "ok", message: "vaccine deleted" });
+        }
+      );
     }
-  );
+  });
 };
